+++ backend/src/main/resources/db/migration/V18.1__add_dropbox_access_token_for_user.sql
